+++ ledger/participant-state/kvutils/src/main/protobuf/com/daml/ledger/participant/state/kvutils/store/events/party_allocation.proto
@@ -1,0 +1,47 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+syntax = "proto3";
+
+package com.daml.ledger.participant.state.kvutils.store.events;
+
+option java_package = "com.daml.ledger.participant.state.kvutils.store.events";
+option java_multiple_files = true;
+option csharp_namespace = "Com.Daml.Ledger.Participant.State.KVUtils.Store.Events";
+
+import "com/daml/ledger/participant/state/kvutils/daml_kvutils.proto";
+
+// An allocation of party name and assignment of a party to a given
+// participant.
+// A party can only be hosted once (this may not be true in more
+// sophisticated ledgers in the future).
+message DamlPartyAllocationEntry {
+  // A unique string scoped to a particular participant.
+  // Implementers are free to select adequate mechanism e.g. UUID or similar.
+  string submission_id = 1;
+
+  // The party name to be allocated.
+  string party = 2;
+
+  // Allocating participant's id.
+  string participant_id = 3;
+
+  // A display name associated with the given party.
+  string display_name = 4;
+}
+
+// A message indicating that party allocation was unsuccessful.
+message DamlPartyAllocationRejectionEntry {
+  // An id selected by the submitter, can be used for matching rejection with
+  // the original upload entry.
+  string submission_id = 1;
+
+  // Uploading participant's id.
+  string participant_id = 2;
+
+  oneof reason {
+    AlreadyExists already_exists = 3;
+    Invalid invalid_name = 4;
+    ParticipantNotAuthorized participant_not_authorized = 5;
+    Duplicate duplicate_submission = 6;
+  }
+}
