+++ ledger/participant-state/kvutils/src/main/protobuf/com/daml/ledger/participant/state/kvutils/store/log_entry.proto
@@ -14,6 +14,7 @@ import "google/protobuf/timestamp.proto";
 import "com/daml/ledger/participant/state/kvutils/daml_kvutils.proto";
 import "com/daml/ledger/participant/state/kvutils/store/events/configuration.proto";
 import "com/daml/ledger/participant/state/kvutils/store/events/package_upload.proto";
+import "com/daml/ledger/participant/state/kvutils/store/events/party_allocation.proto";
 import "com/daml/ledger/participant/state/kvutils/store/events/transaction.proto";
 
 // A log entry for a committed Daml submission.
@@ -49,10 +50,10 @@ message DamlLogEntry {
     DamlConfigurationRejectionEntry configuration_rejection_entry = 7;
 
     // Allocation of a new Daml party and its assignment to a participant.
-    DamlPartyAllocationEntry party_allocation_entry = 8;
+    com.daml.ledger.participant.state.kvutils.store.events.DamlPartyAllocationEntry party_allocation_entry = 8;
 
     // A rejection of party allocation request.
-    DamlPartyAllocationRejectionEntry party_allocation_rejection_entry = 9;
+    com.daml.ledger.participant.state.kvutils.store.events.DamlPartyAllocationRejectionEntry party_allocation_rejection_entry = 9;
 
     // A rejection of a pre-executed submission because of out-of-time-bounds.
     DamlOutOfTimeBoundsEntry out_of_time_bounds_entry = 10;
