+++ ledger/participant-state/kvutils/src/main/protobuf/com/daml/ledger/participant/state/kvutils/wire/submission.proto
@@ -7,9 +7,9 @@ option java_package = "com.daml.ledger.participant.state.kvutils.wire";
 option java_multiple_files = true;
 option csharp_namespace = "Com.Daml.Ledger.Participant.State.KVUtils.Wire";
 
-import "com/daml/ledger/participant/state/kvutils/daml_kvutils.proto";
 import "com/daml/ledger/participant/state/kvutils/store/state.proto";
 import "com/daml/ledger/participant/state/kvutils/store/events/package_upload.proto";
+import "com/daml/ledger/participant/state/kvutils/store/events/party_allocation.proto";
 import "com/daml/ledger/participant/state/kvutils/store/events/transaction.proto";
 import "com/daml/ledger/participant/state/kvutils/wire/configuration_submission.proto";
 
@@ -22,7 +22,7 @@ message DamlSubmission {
     com.daml.ledger.participant.state.kvutils.store.events.DamlTransactionEntry transaction_entry = 2;
     com.daml.ledger.participant.state.kvutils.store.events.DamlPackageUploadEntry package_upload_entry = 3;
     DamlConfigurationSubmission configuration_submission = 4;
-    DamlPartyAllocationEntry party_allocation_entry = 5;
+    com.daml.ledger.participant.state.kvutils.store.events.DamlPartyAllocationEntry party_allocation_entry = 5;
   }
   bytes submission_seed = 6;
 }
