+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/KeyValueSubmission.scala
@@ -6,10 +6,12 @@ package com.daml.ledger.participant.state.kvutils
 import com.daml.daml_lf_dev.DamlLf.Archive
 import com.daml.ledger.configuration.Configuration
 import com.daml.ledger.participant.state.kvutils.Conversions._
-import com.daml.ledger.participant.state.kvutils.DamlKvutils._
 import com.daml.ledger.participant.state.kvutils.store.DamlStateKey
-import com.daml.ledger.participant.state.kvutils.store.events.DamlTransactionEntry
 import com.daml.ledger.participant.state.kvutils.store.events.PackageUpload.DamlPackageUploadEntry
+import com.daml.ledger.participant.state.kvutils.store.events.{
+  DamlPartyAllocationEntry,
+  DamlTransactionEntry,
+}
 import com.daml.ledger.participant.state.kvutils.wire.{DamlConfigurationSubmission, DamlSubmission}
 import com.daml.ledger.participant.state.v2.{SubmitterInfo, TransactionMeta}
 import com.daml.lf.data.Ref
