+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/KVUtilsPartySpec.scala
@@ -3,8 +3,8 @@
 
 package com.daml.ledger.participant.state.kvutils
 
-import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlPartyAllocationRejectionEntry
 import com.daml.ledger.participant.state.kvutils.store.DamlLogEntry
+import com.daml.ledger.participant.state.kvutils.store.events.DamlPartyAllocationRejectionEntry
 import com.daml.logging.LoggingContext
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
