+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/KeyValueConsumptionSpec.scala
@@ -22,6 +22,8 @@ import com.daml.ledger.participant.state.kvutils.store.events.PackageUpload.{
 }
 import com.daml.ledger.participant.state.kvutils.store.events.{
   DamlConfigurationEntry,
+  DamlPartyAllocationEntry,
+  DamlPartyAllocationRejectionEntry,
   DamlSubmitterInfo,
   DamlTransactionEntry,
   DamlTransactionRejectionEntry,
