+++ ledger/participant-state/kvutils/tools/integrity-check/src/test/scala/ledger/participant/state/kvutils/tools/integritycheck/RawPreExecutingCommitStrategySupportSpec.scala
@@ -8,8 +8,8 @@ import java.time.{Duration, Instant, ZoneOffset, ZonedDateTime}
 import com.codahale.metrics.MetricRegistry
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
 import com.daml.ledger.configuration.{Configuration, LedgerTimeModel}
-import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlPartyAllocationEntry
 import com.daml.ledger.participant.state.kvutils.export.{SubmissionInfo, WriteSet}
+import com.daml.ledger.participant.state.kvutils.store.events.DamlPartyAllocationEntry
 import com.daml.ledger.participant.state.kvutils.store.{
   DamlLogEntry,
   DamlStateKey,
